fix: Change startAt field data type to Long (#29)
- Since seconds might be very big integer we are using long

+++ player/src/main/java/com/tpstream/player/TpInitParams.kt
@@ -19,7 +19,7 @@ data class TpInitParams (
     var signature: String? = null,
     var techOverride: Array<String>? = null,
     var isDownloadEnabled: Boolean = false,
-    var startAt: Int = 0
+    var startAt: Long = 0L
 ): Parcelable {
     
     class Builder {
@@ -37,10 +37,10 @@ data class TpInitParams (
         private var signature: String? = null
         private var techOverride: Array<String>? = null
         private var isDownloadEnabled: Boolean = false
-        private var startAt: Int = 0
+        private var startAt: Long = 0L
 
         fun setAutoPlay(autoPlay: Boolean) = apply { this.autoPlay = autoPlay }
-        fun startAt(timeInSeconds: Int) = apply { this.startAt = timeInSeconds }
+        fun startAt(timeInSeconds: Long) = apply { this.startAt = timeInSeconds }
         fun setBufferingGoalMs(bufferingGoalMs: Int) = apply { this.bufferingGoalMs = bufferingGoalMs }
         fun setForceHighestSupportedBitrate(forceHighestSupportedBitrate: Boolean) = apply { this.forceHighestSupportedBitrate = forceHighestSupportedBitrate }
         fun setForceLowestBitrate(forceLowestBitrate: Boolean) = apply { this.forceLowestBitrate = forceLowestBitrate }
